Add dock_smiles_to_protein in Protein class
The addition of this method simplifies the API
as there is no need now to use docking module
for the most basic use cases.
The only argument that must be supplied to this method
is SMILES we want to dock,
so the work that must be done in order to run the benchmark
is minimized.